test: consolidate tenant command test host fixtures
Collapse `TenantAppCommandTestHost`, `TenantRoleCommandTestHost`, and
`TenantPrincipalCommandTestHost` — three ~90%-identical HTTP-mocked test
host fixtures — into a single `TenantCommandTestHost` in
`tests/TALXIS.CLI.Tests/Tenant/`. All tenant-scope CLI command tests
(user, group, app, role) now share the same fake `MicrosoftGraphClient`/
`TenantRoleResolver` wiring instead of each maintaining its own copy of
the fake token service, HTTP handler queue, and connection fixture.
