+++ tests/TALXIS.CLI.Tests/Tenant/App/AppRoleListCliCommandTests.cs
@@ -11,8 +11,8 @@ public sealed class AppRoleListCliCommandTests
     [Fact]
     public async Task RunAsync_ReturnsTenantRoleAndAdminApplicationAssignment()
     {
-        using var host = new TenantAppCommandTestHost(new Queue<Func<HttpRequestMessage, HttpResponseMessage>>([
-            _ => TenantAppCommandTestHost.JsonResponse("""
+        using var host = new TenantCommandTestHost(new Queue<Func<HttpRequestMessage, HttpResponseMessage>>([
+            _ => TenantCommandTestHost.JsonResponse("""
             {
               "value": [
                 {
@@ -23,7 +23,7 @@ public sealed class AppRoleListCliCommandTests
               ]
             }
             """),
-            _ => TenantAppCommandTestHost.JsonResponse("""
+            _ => TenantCommandTestHost.JsonResponse("""
             {
               "value": [
                 {
@@ -35,7 +35,7 @@ public sealed class AppRoleListCliCommandTests
               ]
             }
             """),
-            _ => TenantAppCommandTestHost.JsonResponse("""
+            _ => TenantCommandTestHost.JsonResponse("""
             {
               "value": [
                 {
@@ -48,7 +48,7 @@ public sealed class AppRoleListCliCommandTests
               ]
             }
             """),
-            _ => TenantAppCommandTestHost.JsonResponse("""
+            _ => TenantCommandTestHost.JsonResponse("""
             [
               {
                 "applicationId": "aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"
